iota_core

Module authority

Modules

authority_per_epoch_store

authority_per_epoch_store_pruner

authority_store_pruner

authority_store_tables

authority_store_types

authority_test_utils

backpressure

epoch_start_configuration

shared_object_congestion_tracker

shared_object_version_manager

suggested_gas_price_calculator

test_authority_builder

transaction_deferral

Structs

AuthorityMetrics

Prometheus metrics which can be displayed in Grafana, queried and alerted on

AuthorityState

AuthorityStore

The AuthorityStore manages the state and operations of an authority’s store. It includes a mutex_table to handle concurrent writes to the database and references to various tables stored in AuthorityPerpetualTables. The struct provides mechanisms for initializing and accessing locks, managing objects and transactions, and performing epoch-specific operations. It also includes methods for recovering from crashes, checking IOTA conservation, and handling object markers and states during epoch transitions.

NodeStateDump

ObjDumpFormat

RandomnessRoundReceiver

ResolverWrapper

A wrapper to make Orphan Rule happy

Enums

UpdateType

Constants

SIMULATION_GAS_COIN_VALUE

Gas coin value used in dev-inspect and dry-runs if no gas coin was provided.

Type Aliases

StableSyncAuthoritySigner

a Trait object for Signer that is: