iota_core::authority

Module authority_store_tables

Structs

AuthorityPerpetualTables

AuthorityPerpetualTables contains data that must be preserved from one epoch to the next.

AuthorityPerpetualTablesOptions

Options to apply to every column family of the perpetual DB.

AuthorityPerpetualTablesReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

AuthorityPrunerTables

AuthorityPrunerTablesReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

LiveSetIter

Enums

LiveObject