Module iota_core::authority::authority_test_utils

Functions

build_test_modules_with_dep_addr

certify_shared_obj_transaction_no_execution

certify_transaction

enqueue_all_and_execute_all

execute_certificate_with_execution_error

execute_sequenced_certificate_to_effects

init_certified_transaction

init_certified_transfer_transaction

init_state_validator_with_fullnode

init_state_with_committee

init_state_with_ids

init_state_with_ids_and_expensive_checks

init_state_with_ids_and_versions

init_state_with_object_id

init_state_with_objects

init_state_with_objects_and_committee

init_transfer_transaction

publish_package_on_single_authority

Returns the new package’s ID and the upgrade cap object ref. dep_original_addresses allows us to fill out mappings in the addresses section of the package manifest. These IDs must be the original IDs of names. dep_ids are the IDs of the dependencies of the package, in the latest version (if there were upgrades).

send_and_confirm_transaction

send_and_confirm_transaction_

send_and_confirm_transaction_with_execution_error

send_batch_consensus_no_execution

send_consensus

send_consensus_no_execution

upgrade_package_on_single_authority