iota_core

Module consensus_handler

Structs

ConsensusCommitInfo

Represents the information from the current consensus commit.

ConsensusHandler

ConsensusHandlerInitializer

MysticetiConsensusHandler

Consensus handler used by Mysticeti. Since Mysticeti repo is not yet integrated, we use a channel to receive the consensus output from Mysticeti. During initialization, the sender is passed into Mysticeti which can send consensus output to the channel.

SequencedConsensusTransaction

StarfishConsensusHandler

Consensus handler used by Starfish. During initialization, the sender is passed into Starfish which can send consensus output to the channel.

VerifiedSequencedConsensusTransaction

Enums

SequencedConsensusTransactionKey

SequencedConsensusTransactionKind