iota_core

Module consensus_manager

Modules

mysticeti_manager

starfish_manager

Structs

ConsensusManager

Used by IOTA validator to start consensus protocol for each epoch.

ConsensusManagerMetrics

UpdatableConsensusClient

A ConsensusClient that can be updated internally at any time. This usually happening during epoch change where a client is set after the new consensus is started for the new epoch.

Traits

ConsensusManagerTrait