Module iota_core::execution_cache

Re-exports

pub use passthrough_cache::PassthroughCache;

pub use proxy_cache::ProxyCache;

pub use writeback_cache::WritebackCache;

Modules

metrics

passthrough_cache

proxy_cache

writeback_cache

MemoryCache is a cache for the transaction execution which delays writes to the database until transaction results are certified (i.e. they appear in a certified checkpoint, or an effects cert is observed by a fullnode). The cache also stores committed data in memory in order to serve future reads without hitting the database.

Structs

ExecutionCacheTraitPointers

Enums

ExecutionCacheConfigType

Traits

CheckpointCache

ExecutionCacheAPI

ExecutionCacheCommit

ExecutionCacheReconfigAPI

ExecutionCacheWrite

ObjectCacheRead

StateSyncAPI

TestingAPI

TransactionCacheRead

Functions

build_execution_cache

build_execution_cache_from_env

Should only be used for iota-tool or tests. Nodes must use build_execution_cache which uses the epoch_start_config to prevent cache impl from switching except at epoch boundaries.

choose_execution_cache