iota_core

Module grpc_indexes

Structs

CoinIndexInfo

Coin index value with regulated coin metadata.

CoinIndexKey

GrpcIndexesStore

IndexStoreTablesReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

OwnerIndexInfo

OwnerIndexKey

Hash-based owner index key with fixed-size layout for correct RocksDB byte-order iteration.

Enums

OwnerTypeFilter

Type filter for owner_iter.

Watermark

Checkpoint watermark type

Constants

GRPC_INDEXES_DIR

On-disk directory name for the gRPC indexes store.