iota_core::verify_indexes

Function verify_indexes

pub fn verify_indexes(
    store: &dyn GlobalStateHashStore,
    indexes: Arc<IndexStore>,
) -> Result<()>

This is a very expensive function that verifies some of the secondary indexes. This is done by iterating through the live object set and recalculating these secondary indexes.