iota_core::authority

Module authority_store_tables

Structs

AuthorityPerpetualTables

AuthorityPerpetualTables contains data that must be preserved from one epoch to the next.

AuthorityPerpetualTablesConfigurator

Create config structs for configuring DBMap tables

AuthorityPerpetualTablesOptions

Options to apply to every column family of the perpetual DB.

AuthorityPerpetualTablesReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

AuthorityPrunerTables

AuthorityPrunerTablesConfigurator

Create config structs for configuring DBMap tables

AuthorityPrunerTablesReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

LiveSetIter

Enums

LiveObject