iota_core::authority
Module authority_store_types

Structs

IndirectObjectMetadata

Metadata of stored moved object

StoreMoveObjectV1

Separately stored move object

StoreObjectPair

StoreObjectValue

Forked version of iota_types::object::Object Used for efficient storing of move objects in the database

Enums

StoreData

Forked version of iota_types::object::Data Adds extra enum value IndirectObject, which represents a reference to an object stored separately

StoreMoveObjectWrapper

Enum wrapper for versioning

StoreObjectV1

StoreObjectWrapper

Enum wrapper for versioning

Functions

get_store_object_pair

Type Aliases

ObjectContentDigest

StoreMoveObject

StoreObject