iota_core::execution_cache

Trait ExecutionCacheCommit

pub trait ExecutionCacheCommit: Send + Sync {
    // Required methods
    fn try_commit_transaction_outputs(
        &self,
        epoch: EpochId,
        digests: &[TransactionDigest],
    ) -> IotaResult;
    fn try_persist_transaction(
        &self,
        tx: &VerifiedExecutableTransaction,
    ) -> IotaResult;
    fn approximate_pending_transaction_count(&self) -> u64;

    // Provided methods
    fn commit_transaction_outputs(
        &self,
        epoch: EpochId,
        digests: &[TransactionDigest],
    ) { ... }
    fn persist_transaction(&self, tx: &VerifiedExecutableTransaction) { ... }
}

Required Methods

fn try_commit_transaction_outputs( &self, epoch: EpochId, digests: &[TransactionDigest], ) -> IotaResult

Durably commit the outputs of the given transactions to the database. Will be called by CheckpointExecutor to ensure that transaction outputs are written durably before marking a checkpoint as finalized.

fn try_persist_transaction( &self, tx: &VerifiedExecutableTransaction, ) -> IotaResult

Durably commit a transaction to the database. Used to store any transactions that cannot be reconstructed at start-up by consensus replay. Currently the only case of this is RandomnessStateUpdate.

fn approximate_pending_transaction_count(&self) -> u64

Provided Methods

fn commit_transaction_outputs( &self, epoch: EpochId, digests: &[TransactionDigest], )

Non-fallible version of try_commit_transaction_outputs.

fn persist_transaction(&self, tx: &VerifiedExecutableTransaction)

Non-fallible version of try_persist_transactions.