iota_core::execution_cache

Trait ObjectCacheRead

pub trait ObjectCacheRead: Send + Sync {

    // Required methods
    fn try_get_package_object(
        &self,
        id: &ObjectID,
    ) -> IotaResult<Option<PackageObject>>;
    fn force_reload_system_packages(&self, system_package_ids: &[ObjectID]);
    fn try_get_object(&self, id: &ObjectID) -> IotaResult<Option<Object>>;
    fn try_get_latest_object_ref_or_tombstone(
        &self,
        object_id: ObjectID,
    ) -> IotaResult<Option<ObjectRef>>;
    fn try_get_latest_object_or_tombstone(
        &self,
        object_id: ObjectID,
    ) -> IotaResult<Option<(ObjectKey, ObjectOrTombstone)>>;
    fn try_get_object_by_key(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
    ) -> IotaResult<Option<Object>>;
    fn try_multi_get_objects_by_key(
        &self,
        object_keys: &[ObjectKey],
    ) -> IotaResult<Vec<Option<Object>>>;
    fn try_object_exists_by_key(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
    ) -> IotaResult<bool>;
    fn try_multi_object_exists_by_key(
        &self,
        object_keys: &[ObjectKey],
    ) -> IotaResult<Vec<bool>>;
    fn try_find_object_lt_or_eq_version(
        &self,
        object_id: ObjectID,
        version: SequenceNumber,
    ) -> IotaResult<Option<Object>>;
    fn try_get_lock(
        &self,
        obj_ref: ObjectRef,
        epoch_store: &AuthorityPerEpochStore,
    ) -> IotaResult<ObjectLockStatus>;
    fn _try_get_live_objref(&self, object_id: ObjectID) -> IotaResult<ObjectRef>;
    fn try_check_owned_objects_are_live(
        &self,
        owned_object_refs: &[ObjectRef],
    ) -> IotaResult;
    fn try_get_iota_system_state_object_unsafe(
        &self,
    ) -> IotaResult<IotaSystemState>;
    fn try_get_marker_value(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> IotaResult<Option<MarkerValue>>;
    fn try_get_latest_marker(
        &self,
        object_id: &ObjectID,
        epoch_id: EpochId,
    ) -> IotaResult<Option<(SequenceNumber, MarkerValue)>>;
    fn try_get_highest_pruned_checkpoint(
        &self,
    ) -> IotaResult<CheckpointSequenceNumber>;

    // Provided methods
    fn get_package_object(&self, id: &ObjectID) -> Option<PackageObject> { ... }
    fn get_object(&self, id: &ObjectID) -> Option<Object> { ... }
    fn try_get_objects(
        &self,
        objects: &[ObjectID],
    ) -> IotaResult<Vec<Option<Object>>> { ... }
    fn get_objects(&self, objects: &[ObjectID]) -> Vec<Option<Object>> { ... }
    fn get_latest_object_ref_or_tombstone(
        &self,
        object_id: ObjectID,
    ) -> Option<ObjectRef> { ... }
    fn get_latest_object_or_tombstone(
        &self,
        object_id: ObjectID,
    ) -> Option<(ObjectKey, ObjectOrTombstone)> { ... }
    fn get_object_by_key(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
    ) -> Option<Object> { ... }
    fn multi_get_objects_by_key(
        &self,
        object_keys: &[ObjectKey],
    ) -> Vec<Option<Object>> { ... }
    fn object_exists_by_key(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
    ) -> bool { ... }
    fn multi_object_exists_by_key(&self, object_keys: &[ObjectKey]) -> Vec<bool> { ... }
    fn try_multi_get_objects_with_more_accurate_error_return(
        &self,
        object_refs: &[ObjectRef],
    ) -> Result<Vec<Object>, IotaError> { ... }
    fn multi_get_objects_with_more_accurate_error_return(
        &self,
        object_refs: &[ObjectRef],
    ) -> Vec<Object> { ... }
    fn try_multi_input_objects_available(
        &self,
        keys: &[InputKey],
        receiving_objects: HashSet<InputKey>,
        epoch: EpochId,
    ) -> Result<Vec<bool>, IotaError> { ... }
    fn multi_input_objects_available(
        &self,
        keys: &[InputKey],
        receiving_objects: HashSet<InputKey>,
        epoch: EpochId,
    ) -> Vec<bool> { ... }
    fn find_object_lt_or_eq_version(
        &self,
        object_id: ObjectID,
        version: SequenceNumber,
    ) -> Option<Object> { ... }
    fn get_lock(
        &self,
        obj_ref: ObjectRef,
        epoch_store: &AuthorityPerEpochStore,
    ) -> ObjectLockStatus { ... }
    fn check_owned_objects_are_live(&self, owned_object_refs: &[ObjectRef]) { ... }
    fn get_iota_system_state_object_unsafe(&self) -> IotaSystemState { ... }
    fn get_marker_value(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> Option<MarkerValue> { ... }
    fn get_latest_marker(
        &self,
        object_id: &ObjectID,
        epoch_id: EpochId,
    ) -> Option<(SequenceNumber, MarkerValue)> { ... }
    fn try_get_last_shared_object_deletion_info(
        &self,
        object_id: &ObjectID,
        epoch_id: EpochId,
    ) -> IotaResult<Option<(SequenceNumber, TransactionDigest)>> { ... }
    fn get_last_shared_object_deletion_info(
        &self,
        object_id: &ObjectID,
        epoch_id: EpochId,
    ) -> Option<(SequenceNumber, TransactionDigest)> { ... }
    fn try_get_deleted_shared_object_previous_tx_digest(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> IotaResult<Option<TransactionDigest>> { ... }
    fn get_deleted_shared_object_previous_tx_digest(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> Option<TransactionDigest> { ... }
    fn try_have_received_object_at_version(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> IotaResult<bool> { ... }
    fn have_received_object_at_version(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> bool { ... }
    fn try_have_deleted_owned_object_at_version_or_after(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> IotaResult<bool> { ... }
    fn have_deleted_owned_object_at_version_or_after(
        &self,
        object_id: &ObjectID,
        version: SequenceNumber,
        epoch_id: EpochId,
    ) -> bool { ... }
    fn get_highest_pruned_checkpoint(&self) -> CheckpointSequenceNumber { ... }

}

Required Methods

fn try_get_package_object( &self, id: &ObjectID, ) -> IotaResult<Option<PackageObject>>

fn force_reload_system_packages(&self, system_package_ids: &[ObjectID])

fn try_get_object(&self, id: &ObjectID) -> IotaResult<Option<Object>>

fn try_get_latest_object_ref_or_tombstone( &self, object_id: ObjectID, ) -> IotaResult<Option<ObjectRef>>

fn try_get_latest_object_or_tombstone( &self, object_id: ObjectID, ) -> IotaResult<Option<(ObjectKey, ObjectOrTombstone)>>

fn try_get_object_by_key( &self, object_id: &ObjectID, version: SequenceNumber, ) -> IotaResult<Option<Object>>

fn try_multi_get_objects_by_key( &self, object_keys: &[ObjectKey], ) -> IotaResult<Vec<Option<Object>>>

fn try_object_exists_by_key( &self, object_id: &ObjectID, version: SequenceNumber, ) -> IotaResult<bool>

fn try_multi_object_exists_by_key( &self, object_keys: &[ObjectKey], ) -> IotaResult<Vec<bool>>

fn try_find_object_lt_or_eq_version( &self, object_id: ObjectID, version: SequenceNumber, ) -> IotaResult<Option<Object>>

Return the object with version less then or eq to the provided seq number. This is used by indexer to find the correct version of dynamic field child object. We do not store the version of the child object, but because of lamport timestamp, we know the child must have version number less then or eq to the parent.

fn try_get_lock( &self, obj_ref: ObjectRef, epoch_store: &AuthorityPerEpochStore, ) -> IotaResult<ObjectLockStatus>

fn _try_get_live_objref(&self, object_id: ObjectID) -> IotaResult<ObjectRef>

fn try_check_owned_objects_are_live( &self, owned_object_refs: &[ObjectRef], ) -> IotaResult

fn try_get_iota_system_state_object_unsafe(&self) -> IotaResult<IotaSystemState>

fn try_get_marker_value( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> IotaResult<Option<MarkerValue>>

Get the marker at a specific version

fn try_get_latest_marker( &self, object_id: &ObjectID, epoch_id: EpochId, ) -> IotaResult<Option<(SequenceNumber, MarkerValue)>>

Get the latest marker for a given object.

fn try_get_highest_pruned_checkpoint( &self, ) -> IotaResult<CheckpointSequenceNumber>

Return the watermark for the highest checkpoint for which we’ve pruned objects.

Provided Methods

fn get_package_object(&self, id: &ObjectID) -> Option<PackageObject>

Non-fallible version of try_get_package_object.

fn get_object(&self, id: &ObjectID) -> Option<Object>

Non-fallible version of try_get_object.

fn try_get_objects( &self, objects: &[ObjectID], ) -> IotaResult<Vec<Option<Object>>>

fn get_objects(&self, objects: &[ObjectID]) -> Vec<Option<Object>>

Non-fallible version of try_get_objects.

fn get_latest_object_ref_or_tombstone( &self, object_id: ObjectID, ) -> Option<ObjectRef>

Non-fallible version of try_get_latest_object_ref_or_tombstone.

fn get_latest_object_or_tombstone( &self, object_id: ObjectID, ) -> Option<(ObjectKey, ObjectOrTombstone)>

Non-fallible version of try_get_latest_object_or_tombstone.

fn get_object_by_key( &self, object_id: &ObjectID, version: SequenceNumber, ) -> Option<Object>

Non-fallible version of try_get_object_by_key.

fn multi_get_objects_by_key( &self, object_keys: &[ObjectKey], ) -> Vec<Option<Object>>

Non-fallible version of try_multi_get_objects_by_key.

fn object_exists_by_key( &self, object_id: &ObjectID, version: SequenceNumber, ) -> bool

Non-fallible version of try_object_exists_by_key.

fn multi_object_exists_by_key(&self, object_keys: &[ObjectKey]) -> Vec<bool>

Non-fallible version of try_multi_object_exists_by_key.

fn try_multi_get_objects_with_more_accurate_error_return( &self, object_refs: &[ObjectRef], ) -> Result<Vec<Object>, IotaError>

Load a list of objects from the store by object reference. If they exist in the store, they are returned directly. If any object missing, we try to figure out the best error to return. If the object we are asking is currently locked at a future version, we know this transaction is out-of-date and we return a ObjectVersionUnavailableForConsumption, which indicates this is not retriable. Otherwise, we return a ObjectNotFound error, which indicates this is retriable.

fn multi_get_objects_with_more_accurate_error_return( &self, object_refs: &[ObjectRef], ) -> Vec<Object>

Non-fallible version of try_multi_get_objects_with_more_accurate_error_return.

fn try_multi_input_objects_available( &self, keys: &[InputKey], receiving_objects: HashSet<InputKey>, epoch: EpochId, ) -> Result<Vec<bool>, IotaError>

Used by transaction manager to determine if input objects are ready. Distinct from multi_get_object_by_key because it also consults markers to handle the case where an object will never become available (e.g. because it has been received by some other transaction already).

fn multi_input_objects_available( &self, keys: &[InputKey], receiving_objects: HashSet<InputKey>, epoch: EpochId, ) -> Vec<bool>

Non-fallible version of try_multi_input_objects_available.

fn find_object_lt_or_eq_version( &self, object_id: ObjectID, version: SequenceNumber, ) -> Option<Object>

Non-fallible version of try_find_object_lt_or_eq_version.

fn get_lock( &self, obj_ref: ObjectRef, epoch_store: &AuthorityPerEpochStore, ) -> ObjectLockStatus

Non-fallible version of try_get_lock.

fn check_owned_objects_are_live(&self, owned_object_refs: &[ObjectRef])

Non-fallible version of try_check_owned_objects_are_live.

fn get_iota_system_state_object_unsafe(&self) -> IotaSystemState

Non-fallible version of try_get_iota_system_state_object_unsafe.

fn get_marker_value( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> Option<MarkerValue>

Non-fallible version of try_get_marker_value.

fn get_latest_marker( &self, object_id: &ObjectID, epoch_id: EpochId, ) -> Option<(SequenceNumber, MarkerValue)>

Non-fallible version of try_get_latest_marker.

fn try_get_last_shared_object_deletion_info( &self, object_id: &ObjectID, epoch_id: EpochId, ) -> IotaResult<Option<(SequenceNumber, TransactionDigest)>>

If the shared object was deleted, return deletion info for the current live version

fn get_last_shared_object_deletion_info( &self, object_id: &ObjectID, epoch_id: EpochId, ) -> Option<(SequenceNumber, TransactionDigest)>

Non-fallible version of try_get_last_shared_object_deletion_info.

fn try_get_deleted_shared_object_previous_tx_digest( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> IotaResult<Option<TransactionDigest>>

If the shared object was deleted, return deletion info for the specified version.

fn get_deleted_shared_object_previous_tx_digest( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> Option<TransactionDigest>

Non-fallible version of try_get_deleted_shared_object_previous_tx_digest.

fn try_have_received_object_at_version( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> IotaResult<bool>

fn have_received_object_at_version( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> bool

Non-fallible version of try_have_received_object_at_version.

fn try_have_deleted_owned_object_at_version_or_after( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> IotaResult<bool>

fn have_deleted_owned_object_at_version_or_after( &self, object_id: &ObjectID, version: SequenceNumber, epoch_id: EpochId, ) -> bool

Non-fallible version of try_have_deleted_owned_object_at_version_or_after.

fn get_highest_pruned_checkpoint(&self) -> CheckpointSequenceNumber

Non-fallible version of try_get_highest_pruned_checkpoint.