iota_core::execution_cache

Trait TransactionCacheRead

pub trait TransactionCacheRead: Send + Sync {

    // Required methods
    fn try_multi_get_transaction_blocks(
        &self,
        digests: &[TransactionDigest],
    ) -> IotaResult<Vec<Option<Arc<VerifiedTransaction>>>>;
    fn try_multi_get_executed_effects_digests(
        &self,
        digests: &[TransactionDigest],
    ) -> IotaResult<Vec<Option<TransactionEffectsDigest>>>;
    fn try_multi_get_effects(
        &self,
        digests: &[TransactionEffectsDigest],
    ) -> IotaResult<Vec<Option<TransactionEffects>>>;
    fn try_multi_get_events(
        &self,
        digests: &[TransactionDigest],
    ) -> IotaResult<Vec<Option<TransactionEvents>>>;
    fn try_notify_read_executed_effects_digests<'a>(
        &'a self,
        task_name: &'static str,
        digests: &'a [TransactionDigest],
    ) -> BoxFuture<'a, IotaResult<Vec<TransactionEffectsDigest>>>;

    // Provided methods
    fn multi_get_transaction_blocks(
        &self,
        digests: &[TransactionDigest],
    ) -> Vec<Option<Arc<VerifiedTransaction>>> { ... }
    fn try_get_transaction_block(
        &self,
        digest: &TransactionDigest,
    ) -> IotaResult<Option<Arc<VerifiedTransaction>>> { ... }
    fn get_transaction_block(
        &self,
        digest: &TransactionDigest,
    ) -> Option<Arc<VerifiedTransaction>> { ... }
    fn try_get_transactions_and_serialized_sizes(
        &self,
        digests: &[TransactionDigest],
    ) -> IotaResult<Vec<Option<(VerifiedTransaction, usize)>>> { ... }
    fn get_transactions_and_serialized_sizes(
        &self,
        digests: &[TransactionDigest],
    ) -> Vec<Option<(VerifiedTransaction, usize)>> { ... }
    fn multi_get_executed_effects_digests(
        &self,
        digests: &[TransactionDigest],
    ) -> Vec<Option<TransactionEffectsDigest>> { ... }
    fn try_is_tx_already_executed(
        &self,
        digest: &TransactionDigest,
    ) -> IotaResult<bool> { ... }
    fn is_tx_already_executed(&self, digest: &TransactionDigest) -> bool { ... }
    fn try_multi_get_executed_effects(
        &self,
        digests: &[TransactionDigest],
    ) -> IotaResult<Vec<Option<TransactionEffects>>> { ... }
    fn multi_get_executed_effects(
        &self,
        digests: &[TransactionDigest],
    ) -> Vec<Option<TransactionEffects>> { ... }
    fn try_get_executed_effects(
        &self,
        digest: &TransactionDigest,
    ) -> IotaResult<Option<TransactionEffects>> { ... }
    fn get_executed_effects(
        &self,
        digest: &TransactionDigest,
    ) -> Option<TransactionEffects> { ... }
    fn multi_get_effects(
        &self,
        digests: &[TransactionEffectsDigest],
    ) -> Vec<Option<TransactionEffects>> { ... }
    fn try_get_effects(
        &self,
        digest: &TransactionEffectsDigest,
    ) -> IotaResult<Option<TransactionEffects>> { ... }
    fn get_effects(
        &self,
        digest: &TransactionEffectsDigest,
    ) -> Option<TransactionEffects> { ... }
    fn multi_get_events(
        &self,
        digests: &[TransactionDigest],
    ) -> Vec<Option<TransactionEvents>> { ... }
    fn try_get_events(
        &self,
        digest: &TransactionDigest,
    ) -> IotaResult<Option<TransactionEvents>> { ... }
    fn get_events(
        &self,
        digest: &TransactionDigest,
    ) -> Option<TransactionEvents> { ... }
    fn notify_read_executed_effects_digests<'a>(
        &'a self,
        task_name: &'static str,
        digests: &'a [TransactionDigest],
    ) -> BoxFuture<'a, Vec<TransactionEffectsDigest>> { ... }
    fn try_notify_read_executed_effects<'a>(
        &'a self,
        task_name: &'static str,
        digests: &'a [TransactionDigest],
    ) -> BoxFuture<'a, IotaResult<Vec<TransactionEffects>>> { ... }
    fn notify_read_executed_effects_for_testing<'a>(
        &'a self,
        task_name: &'static str,
        digests: &'a [TransactionDigest],
    ) -> BoxFuture<'a, Vec<TransactionEffects>> { ... }

}

Required Methods

fn try_multi_get_transaction_blocks( &self, digests: &[TransactionDigest], ) -> IotaResult<Vec<Option<Arc<VerifiedTransaction>>>>

fn try_multi_get_executed_effects_digests( &self, digests: &[TransactionDigest], ) -> IotaResult<Vec<Option<TransactionEffectsDigest>>>

fn try_multi_get_effects( &self, digests: &[TransactionEffectsDigest], ) -> IotaResult<Vec<Option<TransactionEffects>>>

fn try_multi_get_events( &self, digests: &[TransactionDigest], ) -> IotaResult<Vec<Option<TransactionEvents>>>

fn try_notify_read_executed_effects_digests<'a>( &'a self, task_name: &'static str, digests: &'a [TransactionDigest], ) -> BoxFuture<'a, IotaResult<Vec<TransactionEffectsDigest>>>

Provided Methods

fn multi_get_transaction_blocks( &self, digests: &[TransactionDigest], ) -> Vec<Option<Arc<VerifiedTransaction>>>

Non-fallible version of try_multi_get_transaction_blocks.

fn try_get_transaction_block( &self, digest: &TransactionDigest, ) -> IotaResult<Option<Arc<VerifiedTransaction>>>

fn get_transaction_block( &self, digest: &TransactionDigest, ) -> Option<Arc<VerifiedTransaction>>

Non-fallible version of try_get_transaction_block.

fn try_get_transactions_and_serialized_sizes( &self, digests: &[TransactionDigest], ) -> IotaResult<Vec<Option<(VerifiedTransaction, usize)>>>

fn get_transactions_and_serialized_sizes( &self, digests: &[TransactionDigest], ) -> Vec<Option<(VerifiedTransaction, usize)>>

Non-fallible version of try_get_transactions_and_serialized_sizes.

fn multi_get_executed_effects_digests( &self, digests: &[TransactionDigest], ) -> Vec<Option<TransactionEffectsDigest>>

Non-fallible version of try_multi_get_executed_effects_digests.

fn try_is_tx_already_executed( &self, digest: &TransactionDigest, ) -> IotaResult<bool>

fn is_tx_already_executed(&self, digest: &TransactionDigest) -> bool

Non-fallible version of try_is_tx_already_executed.

fn try_multi_get_executed_effects( &self, digests: &[TransactionDigest], ) -> IotaResult<Vec<Option<TransactionEffects>>>

fn multi_get_executed_effects( &self, digests: &[TransactionDigest], ) -> Vec<Option<TransactionEffects>>

Non-fallible version of try_multi_get_executed_effects.

fn try_get_executed_effects( &self, digest: &TransactionDigest, ) -> IotaResult<Option<TransactionEffects>>

fn get_executed_effects( &self, digest: &TransactionDigest, ) -> Option<TransactionEffects>

Non-fallible version of try_get_executed_effects.

fn multi_get_effects( &self, digests: &[TransactionEffectsDigest], ) -> Vec<Option<TransactionEffects>>

Non-fallible version of try_multi_get_effects.

fn try_get_effects( &self, digest: &TransactionEffectsDigest, ) -> IotaResult<Option<TransactionEffects>>

fn get_effects( &self, digest: &TransactionEffectsDigest, ) -> Option<TransactionEffects>

Non-fallible version of try_get_effects.

fn multi_get_events( &self, digests: &[TransactionDigest], ) -> Vec<Option<TransactionEvents>>

Non-fallible version of try_multi_get_events.

fn try_get_events( &self, digest: &TransactionDigest, ) -> IotaResult<Option<TransactionEvents>>

fn get_events(&self, digest: &TransactionDigest) -> Option<TransactionEvents>

Non-fallible version of try_get_events.

fn notify_read_executed_effects_digests<'a>( &'a self, task_name: &'static str, digests: &'a [TransactionDigest], ) -> BoxFuture<'a, Vec<TransactionEffectsDigest>>

Non-fallible version of try_notify_read_executed_effects_digests.

fn try_notify_read_executed_effects<'a>( &'a self, task_name: &'static str, digests: &'a [TransactionDigest], ) -> BoxFuture<'a, IotaResult<Vec<TransactionEffects>>>

Wait until the effects of the given transactions are available and return them. WARNING: If calling this on a transaction that could be reverted, you must be sure that this function cannot be called during reconfiguration. The best way to do this is to wrap your future in EpochStore::within_alive_epoch. Holding an ExecutionLockReadGuard would also prevent reconfig from happening while waiting, but this is very dangerous, as it could prevent reconfiguration from ever occurring!

Returns an error if any of the requested effects have been pruned from the database.

fn notify_read_executed_effects_for_testing<'a>( &'a self, task_name: &'static str, digests: &'a [TransactionDigest], ) -> BoxFuture<'a, Vec<TransactionEffects>>

Non-fallible version of try_notify_read_executed_effects.

Panics if any of the requested effects are not found. For paths where effects may have been pruned, use try_notify_read_executed_effects.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".